Refactor LatentSelf into injectable services
